+++ Runtime/ILocatableService.cs
@@ -1,4 +1,0 @@
-namespace DGP.ServiceLocator
-{
-    //public interface ILocatableService {}
-}